--- src/main/kotlin/blackjack/domain/model/hand/BlackJack.kt
@@ -1,3 +1,0 @@
-package blackjack.domain.model.hand
-
-class BlackJack(hands: Hands) : Finished(hands)
