--- src/main/kotlin/blackjack/domain/model/hand/Bust.kt
@@ -1,3 +1,0 @@
-package blackjack.domain.model.hand
-
-class Bust(hands: Hands) : Finished(hands)
