--- src/main/kotlin/blackjack/domain/model/hand/Stay.kt
@@ -1,3 +1,0 @@
-package blackjack.domain.model.hand
-
-class Stay(hands: Hands) : Finished(hands)
