+++ src/main/kotlin/blackjack/domain/model/hand/state/Bust.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.model.hand.state
+
+import blackjack.domain.model.hand.Hands
+
+class Bust(hands: Hands) : Finished(hands)
