+++ src/main/kotlin/blackjack/domain/model/hand/state/Initial.kt
@@ -1,7 +1,8 @@
-package blackjack.domain.model.hand
+package blackjack.domain.model.hand.state
 
 import blackjack.domain.model.Card
 import blackjack.domain.model.Score
+import blackjack.domain.model.hand.Hands
 
 abstract class Initial(protected var hands: Hands) : State {
     override fun cards(): List<Card> = hands.cards.map { it.copy() }
