+++ src/main/kotlin/blackjack/domain/model/hand/state/Playing.kt
@@ -1,4 +1,7 @@
-package blackjack.domain.model.hand
+package blackjack.domain.model.hand.state
+
+import blackjack.domain.model.hand.Hands
+import blackjack.domain.model.hand.strategy.StayStrategy
 
 abstract class Playing(hands: Hands) : Initial(hands) {
     abstract val stayStrategy: StayStrategy
