+++ src/main/kotlin/blackjack/domain/model/hand/state/State.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.model.hand
+package blackjack.domain.model.hand.state
 
 import blackjack.domain.model.Card
 import blackjack.domain.model.Score
