+++ src/main/kotlin/blackjack/domain/model/hand/state/Stay.kt
@@ -1,0 +1,5 @@
+package blackjack.domain.model.hand.state
+
+import blackjack.domain.model.hand.Hands
+
+class Stay(hands: Hands) : Finished(hands)
