+++ src/main/kotlin/blackjack/domain/model/hand/strategy/StayStrategy.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.model.hand
+package blackjack.domain.model.hand.strategy
 
 import blackjack.domain.model.Score
