+++ src/main/kotlin/blackjack/domain/model/playing/PlayingDealer.kt
@@ -1,10 +1,10 @@
 package blackjack.domain.model.playing
 
 import blackjack.domain.model.Card
-import blackjack.domain.model.hand.DealerStay
 import blackjack.domain.model.hand.Hands
-import blackjack.domain.model.hand.Hit
-import blackjack.domain.model.hand.State
+import blackjack.domain.model.hand.state.Hit
+import blackjack.domain.model.hand.state.State
+import blackjack.domain.model.hand.strategy.DealerStay
 
 class PlayingDealer(override var handsState: State, override val name: String = DEALER_NAME) : PlayingParticipant() {
     constructor(vararg card: Card) : this(Hit(DealerStay(), Hands(card.toList())))
