+++ src/main/kotlin/blackjack/domain/model/playing/PlayingPlayer.kt
@@ -2,9 +2,9 @@ package blackjack.domain.model.playing
 
 import blackjack.domain.model.Card
 import blackjack.domain.model.hand.Hands
-import blackjack.domain.model.hand.Hit
-import blackjack.domain.model.hand.PlayerStay
-import blackjack.domain.model.hand.State
+import blackjack.domain.model.hand.state.Hit
+import blackjack.domain.model.hand.state.State
+import blackjack.domain.model.hand.strategy.PlayerStay
 
 class PlayingPlayer(override var handsState: State, override val name: String) : PlayingParticipant() {
     constructor(name: String, vararg card: Card) : this(Hit(PlayerStay(), Hands(card.toList())), name)
